Linux perf support - fix debug info.
This fixes support for debug info in perf. Thanks to Stephane Eranian for
identifying the problem - debug info event has to be emitted before the
code load event. It also seems that perf does not yet support the shorthand
for repeated source files in the debug info entry ("\xff\0"), so I changed
it to always write the script name.

Review URL: https://codereview.chromium.org/1843563002

Cr-Commit-Position: refs/heads/master@{#35092}