[turbofan] Fixes to validator
If we have 2 phis with the exact same operand list, and the first phi is
used before the second one, via the operand incoming to the block
that defines the phi, and the second one's operand is defined (via a
parallel move) after the use, then the original operand will be assigned
to the first phi. This will lead to a spurious validation error.

To fix this, we look at the original pending assessment.

Review URL: https://codereview.chromium.org/1895013003

Cr-Commit-Position: refs/heads/master@{#35601}